Refactor: move AS reflection tests to more common architecture test
I want to add additional tests relying on reflection to verify correct adoption of the general CARP architecture. Rather than introducing this is yet another base class in each submodule, I will move such "architecture tests" to a single class per submodule.
