+++ carp.data.core/src/jvmTest/kotlin/dk/cachet/carp/data/DataArchitectureTest.kt
@@ -1,0 +1,6 @@
+package dk.cachet.carp.data
+
+import dk.cachet.carp.common.test.SubsystemArchitectureTest
+
+
+class DataArchitectureTest : SubsystemArchitectureTest( DataArchitectureTest::class.java.`package`.name )
