--- carp.data.core/src/jvmTest/kotlin/dk/cachet/carp/data/application/DataStreamServicesReflectionTest.kt
@@ -1,7 +1,0 @@
-package dk.cachet.carp.data.application
-
-import dk.cachet.carp.common.test.application.ApplicationServicesReflectionTest
-
-
-class DataStreamServicesReflectionTest :
-    ApplicationServicesReflectionTest( DataStreamServicesReflectionTest::class.java.`package`.name )
