+++ carp.deployments.core/src/jvmTest/kotlin/dk/cachet/carp/deployments/DeploymentsArchitectureTest.kt
@@ -1,0 +1,6 @@
+package dk.cachet.carp.deployments
+
+import dk.cachet.carp.common.test.SubsystemArchitectureTest
+
+
+class DeploymentsArchitectureTest : SubsystemArchitectureTest( DeploymentsArchitectureTest::class.java.`package`.name )
