--- carp.deployments.core/src/jvmTest/kotlin/dk/cachet/carp/deployments/application/DeploymentsServicesReflectionTest.kt
@@ -1,7 +1,0 @@
-package dk.cachet.carp.deployments.application
-
-import dk.cachet.carp.common.test.application.ApplicationServicesReflectionTest
-
-
-class DeploymentsServicesReflectionTest :
-    ApplicationServicesReflectionTest( DeploymentsServicesReflectionTest::class.java.`package`.name )
