+++ carp.protocols.core/src/jvmTest/kotlin/dk/cachet/carp/protocols/ProtocolsArchitectureTest.kt
@@ -1,0 +1,6 @@
+package dk.cachet.carp.protocols
+
+import dk.cachet.carp.common.test.SubsystemArchitectureTest
+
+
+class ProtocolsArchitectureTest : SubsystemArchitectureTest( ProtocolsArchitectureTest::class.java.`package`.name )
