--- carp.protocols.core/src/jvmTest/kotlin/dk/cachet/carp/protocols/application/ProtocolsServicesReflectionTest.kt
@@ -1,7 +1,0 @@
-package dk.cachet.carp.protocols.application
-
-import dk.cachet.carp.common.test.application.ApplicationServicesReflectionTest
-
-
-class ProtocolsServicesReflectionTest :
-    ApplicationServicesReflectionTest( ProtocolsServicesReflectionTest::class.java.`package`.name )
