+++ carp.studies.core/src/jvmTest/kotlin/dk/cachet/carp/studies/StudiesArchitectureTest.kt
@@ -1,0 +1,6 @@
+package dk.cachet.carp.studies
+
+import dk.cachet.carp.common.test.SubsystemArchitectureTest
+
+
+class StudiesArchitectureTest : SubsystemArchitectureTest( StudiesArchitectureTest::class.java.`package`.name )
