--- carp.studies.core/src/jvmTest/kotlin/dk/cachet/carp/studies/application/StudiesServicesReflectionTest.kt
@@ -1,7 +1,0 @@
-package dk.cachet.carp.studies.application
-
-import dk.cachet.carp.common.test.application.ApplicationServicesReflectionTest
-
-
-class StudiesServicesReflectionTest :
-    ApplicationServicesReflectionTest( StudiesServicesReflectionTest::class.java.`package`.name )
